8 implement spring security (#15)
* Added alooooooooooooot of stuff... including rbac with jwt and login functionality
(missing thymeleaf template for login though)

* Adjusted GET user methods in UserController & UserService

* Created method 'registerUser' that can register user with ROLE_USER or ROLE_ADMIN

* Added new method & request 'logout'

* Upgraded deprecated methods in SecurityConfig.java (WebSecurityConfigurerAdapter extension)

* Added @SuppressWarnings annotation on AuthTokenFilter.java

* Added CorsConfigurationSource to SecurityConfig.java so that frontend can fetch from API.
Also added another signup method for user registration where ROLE_USER is default

+++ src/main/java/projectwork/backend/model/Role.java
@@ -3,10 +3,9 @@ package projectwork.backend.model;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import projectwork.backend.enums.ERole;
+import projectwork.backend.model.enums.ERole;
 
 import javax.persistence.*;
-import java.util.Set;
 
 @Entity
 @Table(name = "roles")
@@ -18,7 +17,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
     private ERole role;
